Move PubSubChannel to lib/util
