+++ apps/src/lib/util/PubSubService.js
@@ -3,7 +3,7 @@
  */
 
 import Pusher from 'pusher-js';
-var PubSubChannel = require('../../netsim/PubSubChannel');
+var PubSubChannel = require('./PubSubChannel');
 
 /**
  * JavaScript interface for a publish/subscribe service provider.
